refactor: 버튼 이동 데이터 - 경로 변경
