+++ src/features/admin/ui/table-container.tsx
@@ -1,7 +1,13 @@
 import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
-import { Button, getDynamicPath, ItemTable, RouterPath } from '@/shared';
+import {
+  Button,
+  getDynamicPath,
+  ItemTable,
+  NAVIGATE_BUTTONS_DATA,
+  RouterPath,
+} from '@/shared';
 
 import { ITEM_BODY, ITEM_HEADERS } from '../model';
 
@@ -24,7 +30,7 @@ export const TableContainer = () => {
 
   return (
     <>
-      <p>해달 동아리 물품 대여 상태입니다.</p>
+      <p>{NAVIGATE_BUTTONS_DATA[3].description}</p>
       <div className='w-full px-3'>
         <div className='flex w-full justify-between'>
           <div className='flex gap-2'>
