--- src/features/main/model/data/index.ts
@@ -1,1 +1,0 @@
-export * from './navigate-buttons-data';
