+++ src/pages/item-list/ItemListPage.tsx
@@ -1,12 +1,11 @@
 import { useNavigate } from 'react-router-dom';
 
 import { ItemListDescriptionSection, ItemListSection } from '@/features';
-import { NAVIGATE_BUTTONS_DATA } from '@/features/main/model';
 
 // Todo: DescriptionSection 스타일을 컴포넌트화할 필요
 // 같은 절대 경로에서 import하는 모듈들 이름 재설정 (admin 내 모듈들)
 
-import { Button } from '@/shared';
+import { Button, NAVIGATE_BUTTONS_DATA } from '@/shared';
 
 import { LogoContainer } from '@/widgets';
