--- src/shared/constants/description-data.ts
@@ -1,18 +1,0 @@
-export const DESCRIPTION_DATA = [
-  {
-    description: '해달 동아리방 물품 목록입니다.',
-    key: 'search',
-  },
-  {
-    description: '',
-    key: 'write',
-  },
-  {
-    description: '님이 대여한 물품 목록입니다.',
-    key: 'my-rent',
-  },
-  {
-    description: '해달 동아리방 물품을 관리합니다.',
-    key: 'admin',
-  },
-] as const;
