+++ src/shared/constants/index.ts
@@ -1,3 +1,3 @@
 export * from './router-path';
 export * from './dummy-token';
-export * from './description-data';
+export * from './navigate-buttons-data';
